requirements.txt
PyYAML
amulet-core